refactor: split CreateViewTemplates into nested route sections
