+++ src/views/CreateNew/sections/StartFromTemplateSection.vue
@@ -1,0 +1,19 @@
+<script setup>
+  // Start from Template section - placeholder for future implementation
+</script>
+
+<template>
+  <div class="flex-1 flex flex-col gap-5 min-w-0">
+    <!-- Start from Template Content - To be implemented -->
+    <div class="flex-1 p-6 bg-[var(--surface-50)] rounded-md border surface-border">
+      <div class="flex flex-col items-center justify-center h-full min-h-[300px] text-center">
+        <i class="pi pi-file text-4xl text-color-secondary mb-4"></i>
+        <h3 class="text-sm font-semibold text-color mb-2">Start from Template</h3>
+        <p class="text-xs text-color-secondary max-w-md">
+          Choose from a variety of pre-configured templates to quickly set up your project. This
+          section will be available soon.
+        </p>
+      </div>
+    </div>
+  </div>
+</template>
